+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
-- Fix action/function return type structure for complex and primitive types
+- Change action/function return type value representation for complex and primitive types
+- Introduce proxy options 'returnComplexNested' and 'returnPrimitivePlain' to keep previous action/function return value representation
 
 ## Version 1.6.3 - 2021-05-26
 
@@ -90,7 +91,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Support action/function array parameter types
-- Introduce proxy option for body parser size limit
+- Introduce proxy option 'bodyParserLimit' for body parser size limit
 
 ## Version 1.5.3 - 2021-01-12